Many changes mostly to get the checkout flow back to working
- Moved the checkout interstitial to the cart app (seemed more appropriate there, given that has frontend stuff in it)
- Finished up adapting APIs to work for UE
- Added the migration to pull the dependency on system from the basket (they're unified, baskets aren't per system)
- Some more updates on the cart test mule to make it more functional
- Added a bunch of CyberSource constants that should go into ol-django at some point

At this point checkout seems to be working up to the point where the callback gets hit; there's some bugs in that still. This is a no-verify commit because the API tests aren't shored up yet.
